Move compact tests into the module-specific test files
Deletes compact_test.py and relocates its tests to follow the one-_test.py-per-module
pattern: the active-DOF bookkeeping tests (wake state driving island.update_active_dofs)
go to sleep_test.py as ActiveDofTest, and the compacted smooth/constrained solve tests go
to solver_test.py as CompactSolverTest.
